mom-safety-monitor/requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
paho-mqtt==1.5.1
